@@ -4,22 +4,22 @@
   "dependencies": {
     "@rails/webpacker": "5.2.1",
     "@storybook/cli": "^6.1.15",
-    "axios": "^0.21.1",
+    "@types/axios": "^0.14.0",
+    "@vue/compiler-sfc": "^3.0.11",
     "eslint": "^7.20.0",
     "eslint-plugin-vue": "^7.5.0",
-    "vee-validate": "^3.4.5",
-    "vue": "^2.6.12",
-    "vue-loader": "^15.9.6",
+    "vue": "^3.0.11",
+    "vue-loader": "^16.2.0",
     "vue-star-rating": "^1.7.0",
     "vue-template-compiler": "^2.6.12",
     "vuex": "^3.6.2"
   },
   "devDependencies": {
     "@babel/core": "^7.12.10",
-    "@storybook/addon-actions": "^6.1.15",
-    "@storybook/addon-essentials": "^6.1.15",
-    "@storybook/addon-links": "^6.1.15",
-    "@storybook/vue": "^6.1.15",
+    "@storybook/addon-actions": "^6.3.0-beta.4",
+    "@storybook/addon-essentials": "^6.3.0-beta.4",
+    "@storybook/addon-links": "^6.3.0-beta.4",
+    "@storybook/vue3": "^6.3.0-beta.4",
     "babel-loader": "^8.2.2",
     "chromatic": "^5.6.1",
     "webpack-dev-server": "^3.11.1"
