@@ -4,12 +4,13 @@ module.exports = {
   "stories": [
     "../stories/**/*.stories.mdx",
     "../stories/**/*.stories.@(js|jsx|ts|tsx)",
-    "../app/javascript/**/*.stories.@(js|jsx|ts|tsx)"
+    "../app/javascript/**/*.stories.@(js|jsx|ts|tsx)",
   ],
   "addons": [
     "@storybook/addon-links",
     "@storybook/addon-essentials"
   ],
+
   webpackFinal: async (config) => {
     config.resolve.alias['@axios'] = path.resolve(__dirname, '../app/javascript/packs/initializers/axios.js');
     return config;
