@@ -1,9 +1,18 @@
 const { environment } = require('@rails/webpacker')
 const { VueLoaderPlugin } = require('vue-loader')
-const vue = require('./loaders/vue')
+const vue = require("./loaders/vue");
 const path = require('path')
-
 const webpack = require('webpack')
+const { DefinePlugin } = require('webpack')
+
+environment.plugins.prepend('VueLoaderPlugin', new VueLoaderPlugin())
+environment.loaders.prepend('vue', vue)
+
+environment.config.resolve.alias = {
+  '@axios': path.resolve(__dirname, '../../app/javascript/packs/initializers/axios.js'),
+  '@': path.resolve(__dirname, "..", "..", "app/javascript/src")
+}
+
 environment.plugins.prepend(
   'Provide',
   new webpack.ProvidePlugin({
@@ -12,11 +21,15 @@ environment.plugins.prepend(
     Popper: 'popper.js'
   })
 )
-environment.plugins.prepend('VueLoaderPlugin', new VueLoaderPlugin())
-environment.loaders.prepend('vue', vue)
-environment.config.resolve.alias = {
-  'vue$': 'vue/dist/vue.esm.js',
-  '@axios': path.resolve(__dirname, '../../app/javascript/packs/initializers/axios.js')
-}
+
+// NOE: 実行時にWarningが出ないための設定
+environment.plugins.prepend(
+  'Define',
+  new DefinePlugin({
+    __VUE_OPTIONS_API__: false,
+    // or __VUE_OPTIONS_API__: true,
+    __VUE_PROD_DEVTOOLS__: false
+  })
+)
 
 module.exports = environment
