@@ -6,13 +6,19 @@ export default {
   argTypes: {
     backgroundColor: { control: 'color' },
     size: { control: { type: 'select', options: ['small', 'medium', 'large'] } },
+    onClick: {},
   },
 };
 
-const Template = (args, { argTypes }) => ({
-  props: Object.keys(argTypes),
+const Template = (args) => ({
+  // Components used in your story `template` are defined in the `components` object
   components: { MyButton },
-  template: '<my-button @onClick="onClick" v-bind="$props" />',
+  // The story's `args` need to be mapped into the template through the `setup()` method
+  setup() {
+    return { args };
+  },
+  // And then the `args` are bound to your component with `v-bind="args"`
+  template: '<my-button v-bind="args" />',
 });
 
 export const Primary = Template.bind({});
