@@ -4,6 +4,7 @@
 
 <script>
 import './button.css';
+import { reactive, computed } from 'vue';
 
 export default {
   name: 'my-button',
@@ -19,7 +20,6 @@ export default {
     },
     size: {
       type: String,
-      default: 'medium',
       validator: function (value) {
         return ['small', 'medium', 'large'].indexOf(value) !== -1;
       },
@@ -29,26 +29,24 @@ export default {
     },
   },
 
-  computed: {
-    classes() {
-      return {
-        'storybook-button': true,
-        'storybook-button--primary': this.primary,
-        'storybook-button--secondary': !this.primary,
-        [`storybook-button--${this.size}`]: true,
-      };
-    },
-    style() {
-      return {
-        backgroundColor: this.backgroundColor,
-      };
-    },
-  },
+  emits: ['click'],
 
-  methods: {
-    onClick() {
-      this.$emit('onClick');
-    },
+  setup(props, { emit }) {
+    props = reactive(props);
+    return {
+      classes: computed(() => ({
+        'storybook-button': true,
+        'storybook-button--primary': props.primary,
+        'storybook-button--secondary': !props.primary,
+        [`storybook-button--${props.size || 'medium'}`]: true,
+      })),
+      style: computed(() => ({
+        backgroundColor: props.backgroundColor,
+      })),
+      onClick() {
+        emit('click');
+      }
+    }
   },
 };
 </script>
