@@ -21,9 +21,9 @@
         <h1>Acme</h1>
       </div>
       <div>
-        <my-button size="small" @onClick="onLogout" label="Log out" v-if="user" />
-        <my-button size="small" @onClick="onLogin" label="Log in" v-if="!user" />
-        <my-button primary size="small" @onClick="onCreateAccount" label="Sign up" v-if="!user" />
+        <my-button size="small" @click="$emit('logout')" label="Log out" v-if="user" />
+        <my-button size="small" @click="$emit('login')" label="Log in" v-if="!user" />
+        <my-button primary size="small" @click="$emit('createAccount')" label="Sign up" v-if="!user" />
       </div>
     </div>
   </header>
@@ -44,16 +44,6 @@ export default {
     },
   },
 
-  methods: {
-    onLogin() {
-      this.$emit('onLogin');
-    },
-    onLogout() {
-      this.$emit('onLogout');
-    },
-    onCreateAccount() {
-      this.$emit('onCreateAccount');
-    },
-  },
+  emits: ['login', 'logout', 'createAccount'],
 };
 </script>
