@@ -1,4 +1,4 @@
-import MyPage from './Page';
+import MyPage from './Page.vue';
 import * as HeaderStories from './Header.stories';
 
 export default {
@@ -6,11 +6,16 @@ export default {
   component: MyPage,
 };
 
-const Template = (args, { argTypes }) => ({
-  props: Object.keys(argTypes),
+const Template = (args) => ({
+  // Components used in your story `template` are defined in the `components` object
   components: { MyPage },
-  template:
-    '<my-page :user="user" @onLogin="onLogin" @onLogout="onLogout" @onCreateAccount="onCreateAccount" />',
+  // The story's `args` need to be mapped into the template through the `setup()` method
+  setup() {
+    // Story args can be mapped to keys in the returned object
+    return { user: args.user };
+  },
+  // Then, those values can be accessed directly in the template
+  template: '<my-page :user="user" />',
 });
 
 export const LoggedIn = Template.bind({});
